+++ src/codesyncd/utils.ts
@@ -18,42 +18,38 @@ import { uploadFileToServer } from '../utils/upload_utils';
 import { CodeSyncLogger } from '../logger';
 import { generateSettings } from "../settings";
 import { pathUtils } from "../utils/path_utils";
-import { checkSubDir, getActiveUsers, isRepoActive, readYML } from '../utils/common';
+import { checkSubDir, getActiveUsers, isRepoActive, readFile, readYML } from '../utils/common';
 import { getPlanLimitReached } from '../utils/pricing_utils';
 import { CodeSyncState, CODESYNC_STATES } from '../utils/state_utils';
 
 
 export const isValidDiff = (diffData: IDiff) => {
 	const missingKeys = REQUIRED_DIFF_KEYS.filter(key => !(key in diffData));
-	if (missingKeys.length) { return false; }
+	if (missingKeys.length) return false;
 	const isRename = diffData.is_rename;
 	const isDirRename = diffData.is_dir_rename;
 	const diff = diffData.diff;
-	if (diff && diff.length > DIFF_SIZE_LIMIT) { return false; }
+	if (diff && diff.length > DIFF_SIZE_LIMIT) return false;
 	if (isRename || isDirRename) {
-		if (!diff) { return false; }
+		if (!diff) return false;
 		let diffJSON = {};
 		diffJSON = yaml.load(diff);
-		if (typeof diffJSON !== "object") {
-			return false;
-		}
-		if (isRename && isDirRename) {
-			return false;
-		}
+		if (typeof diffJSON !== "object") return false;
+		if (isRename && isDirRename) return false;
 		if (isRename) {
 			const missingRenameKeys = REQUIRED_FILE_RENAME_DIFF_KEYS.filter(key => !(key in diffJSON));
-			if (missingRenameKeys.length) { return false; }
+			if (missingRenameKeys.length) return false;
 		}
 		if (isDirRename) {
 			const missingDirRenameKeys = REQUIRED_DIR_RENAME_DIFF_KEYS.filter(key => !(key in diffJSON));
-			if (missingDirRenameKeys.length) { return false; }
+			if (missingDirRenameKeys.length) return false;
 		}
 	}
 	return true;
 };
 
 export const handleNewFileUpload = async (accessToken: string, repoPath: string, branch: string, addedAt: string,
-											relPath: string, repoId: number, configJSON: any) => {
+											relPath: string, repoId: number, configJSON: any, deleteDiff=true) => {
 	/*
 		Uploads new file to server and adds it in config
 		Ignores if file is not present in .originals repo
@@ -81,7 +77,7 @@ export const handleNewFileUpload = async (accessToken: string, repoPath: string,
 		CodeSyncLogger.error(`Error uploading file: ${response.error}`);
 		return {
 			uploaded: false,
-			deleteDiff: response.statusCode === 404 ? true : false,
+			deleteDiff: response.statusCode === 404,
 			config: configJSON
 		};
 	}
@@ -96,7 +92,7 @@ export const handleNewFileUpload = async (accessToken: string, repoPath: string,
 
 	return {
 		uploaded: true,
-		deleteDiff: true,
+		deleteDiff: deleteDiff,
 		config: configJSON
 	};
 };
@@ -131,7 +127,7 @@ export const getDIffForDeletedFile = (repoPath: string, branch: string, relPath:
 		cleanUpDeleteDiff(repoPath, branch, relPath, configJSON);
 		return diff;
 	}
-	const shadowText = fs.readFileSync(shadowPath, "utf8");
+	const shadowText = readFile(shadowPath);
 	const dmp = new diff_match_patch();
 	const patches = dmp.patch_make(shadowText, "");
 	diff = dmp.patch_toText(patches);
@@ -139,46 +135,6 @@ export const getDIffForDeletedFile = (repoPath: string, branch: string, relPath:
 	return diff;
 };
 
-export const similarity = (s1: string, s2: string) => {
-	let longer = s1;
-	let shorter = s2;
-	if (s1.length < s2.length) {
-		longer = s2;
-		shorter = s1;
-	}
-	const longerLength = longer.length;
-	if (longerLength == 0) {
-		return 1.0;
-	}
-	return (longerLength - editDistance(longer, shorter)) / longerLength;
-};
-
-const editDistance = (s1: string, s2: string) => {
-	s1 = s1.toLowerCase();
-	s2 = s2.toLowerCase();
-
-	const costs: number[] = [];
-	for (let i = 0; i <= s1.length; i++) {
-		let lastValue = i;
-		for (let j = 0; j <= s2.length; j++) {
-			if (i == 0)
-				costs[j] = j;
-			else {
-				if (j > 0) {
-					let newValue = costs[j - 1];
-					if (s1.charAt(i - 1) != s2.charAt(j - 1))
-						newValue = Math.min(Math.min(newValue, lastValue),
-							costs[j]) + 1;
-					costs[j - 1] = lastValue;
-					lastValue = newValue;
-				}
-			}
-		}
-		if (i > 0)
-			costs[s2.length] = lastValue;
-	}
-	return costs[s2.length];
-};
 
 export class statusBarMsgs {
 	/*