+++ src/codesyncd/websocket/socket_client.ts
@@ -7,12 +7,13 @@ import {recallDaemon} from "../codesyncd";
 import {SocketEvents} from "./socket_events";
 import {
     CONNECTION_ERROR_MESSAGE,
-    DIFF_SOURCE,
     SOCKET_CONNECT_ERROR_CODES,
     SOCKET_ERRORS,
-    API_ROUTES
+    API_ROUTES,
+    RETRY_WEBSOCKET_CONNECTION_AFTER
 } from "../../constants";
 import { getPlanLimitReached } from "../../utils/pricing_utils";
+import { CodeSyncState, CODESYNC_STATES } from "../../utils/state_utils";
 
 let errorCount = 0;
 
@@ -31,6 +32,8 @@ export class SocketClient {
     }
 
     resetGlobals = () => {
+        // Set time when connection is errored out
+        CodeSyncState.set(CODESYNC_STATES.WEBSOCKET_ERROR_OCCURRED_AT, new Date().getTime());
         try {
             this.client.abort();
         } catch (e) {
@@ -46,6 +49,9 @@ export class SocketClient {
         const { planLimitReached, canRetry } = getPlanLimitReached();
 		if (planLimitReached && !canRetry) return recallDaemon(this.statusBarItem);
 
+        const errorOccurredAt = CodeSyncState.get(CODESYNC_STATES.WEBSOCKET_ERROR_OCCURRED_AT);
+		const canConnect = !errorOccurredAt || (new Date().getTime() - errorOccurredAt) > RETRY_WEBSOCKET_CONNECTION_AFTER;
+		if (!canConnect) return recallDaemon(this.statusBarItem);
         if (!this.client) {
             this.client = new client();
             (global as any).client = this.client;
@@ -78,7 +84,7 @@ export class SocketClient {
             that.registerConnectionEvents(connection);
         });
 
-        let url = `${API_ROUTES.DIFFS_WEBSOCKET}?token=${this.accessToken}&source=${DIFF_SOURCE}`;
+        let url = `${API_ROUTES.DIFFS_WEBSOCKET}&token=${this.accessToken}`;
         if (!canSendDiffs) {
             url += '&auth_only=1';
         }