+++ tests/test_init/utils.test.js
@@ -14,9 +14,10 @@ import {
     TEST_REPO_RESPONSE,
     TEST_USER,
     randomBaseRepoPath,
-    randomRepoPath, getConfigFilePath, getSyncIgnoreFilePath, getUserFilePath, getSeqTokenFilePath, Config, addUser, waitFor
+    randomRepoPath, getConfigFilePath, getSyncIgnoreFilePath, getUserFilePath, getSeqTokenFilePath, Config, addUser, waitFor,
+    writeTestRepoFiles
 } from "../helpers/helpers";
-import {API_ROUTES, DEFAULT_BRANCH, DIFF_SOURCE, NOTIFICATION, SYNCIGNORE} from "../../src/constants";
+import {API_ROUTES, DEFAULT_BRANCH, VSCODE, NOTIFICATION, SYNCIGNORE} from "../../src/constants";
 import {readYML} from "../../src/utils/common";
 import fetchMock from "jest-fetch-mock";
 import {isBinaryFileSync} from "isbinaryfile";
@@ -57,17 +58,16 @@ describe("getSyncablePaths",  () => {
         fs.writeFileSync(syncIgnorePath, SYNC_IGNORE_DATA+"\nignore.js");
         const initUtilsObj = new initUtils(repoPath);
         const paths = initUtilsObj.getSyncablePaths();
-
         // 1 is .syncignore, other is file.js
         expect(paths).toHaveLength(2);
-        expect(paths[0].rel_path).toStrictEqual(SYNCIGNORE);
+        expect(paths[0].rel_path).toStrictEqual("file.js");
         expect(paths[0].is_binary).toBe(false);
-        expect(paths[0].file_path).toStrictEqual(syncIgnorePath);
-        expect(paths[0].size).toBeTruthy();
-        expect(paths[1].rel_path).toStrictEqual("file.js");
+        expect(paths[0].file_path).toStrictEqual(filePath);
+        expect(paths[0].size).toStrictEqual(0);
+        expect(paths[1].rel_path).toStrictEqual(SYNCIGNORE);
         expect(paths[1].is_binary).toBe(false);
-        expect(paths[1].file_path).toStrictEqual(filePath);
-        expect(paths[1].size).toStrictEqual(0);
+        expect(paths[1].file_path).toStrictEqual(syncIgnorePath);
+        expect(paths[1].size).toBeTruthy();
     });
 });
 
@@ -121,6 +121,7 @@ describe("copyFilesTo",  () => {
 
 describe("saveIamUser",  () => {
     const baseRepoPath = randomBaseRepoPath();
+    const repoPath = randomRepoPath();
     const userFilePath = getUserFilePath(baseRepoPath);
     const userFileData = {};
     userFileData[TEST_USER.email] = {
@@ -139,7 +140,7 @@ describe("saveIamUser",  () => {
     });
 
     test("With no user.yml",  () => {
-        const initUtilsObj = new initUtils();
+        const initUtilsObj = new initUtils(repoPath);
         initUtilsObj.saveIamUser(TEST_USER);
         expect(fs.existsSync(userFilePath)).toBe(true);
         const users = readYML(userFilePath);
@@ -149,7 +150,7 @@ describe("saveIamUser",  () => {
 
     test("With no active user.yml",  () => {
         addUser(baseRepoPath, false);
-        const initUtilsObj = new initUtils();
+        const initUtilsObj = new initUtils(repoPath);
         initUtilsObj.saveIamUser(TEST_USER);
         expect(fs.existsSync(userFilePath)).toBe(true);
         const users = readYML(userFilePath);
@@ -161,7 +162,7 @@ describe("saveIamUser",  () => {
         fs.writeFileSync(userFilePath, yaml.safeDump(userFileData));
         const testUser = Object.assign({}, TEST_USER);
         testUser.email = ANOTHER_TEST_EMAIL;
-        const initUtilsObj = new initUtils();
+        const initUtilsObj = new initUtils(repoPath);
         initUtilsObj.saveIamUser(testUser);
         expect(fs.existsSync(userFilePath)).toBe(true);
         const users = readYML(userFilePath);
@@ -176,7 +177,7 @@ describe("saveIamUser",  () => {
         fs.writeFileSync(userFilePath, yaml.safeDump(userFileData));
         const testUser = Object.assign({}, TEST_USER);
         testUser.email = TEST_EMAIL;
-        const initUtilsObj = new initUtils();
+        const initUtilsObj = new initUtils(repoPath);
         initUtilsObj.saveIamUser(testUser);
         expect(fs.existsSync(userFilePath)).toBe(true);
         const users = readYML(userFilePath);
@@ -188,6 +189,7 @@ describe("saveIamUser",  () => {
 
 describe("saveSequenceTokenFile",  () => {
     const baseRepoPath = randomBaseRepoPath();
+    const repoPath = randomRepoPath();
     const sequenceTokenFilePath = getSeqTokenFilePath(baseRepoPath);
     const sequenceTokenFileData = {};
     sequenceTokenFileData[TEST_EMAIL] = "";
@@ -203,7 +205,7 @@ describe("saveSequenceTokenFile",  () => {
     });
 
     test("With no sequence_token.yml",  () => {
-        const initUtilsObj = new initUtils();
+        const initUtilsObj = new initUtils(repoPath);
         initUtilsObj.saveSequenceTokenFile(TEST_EMAIL);
         expect(fs.existsSync(sequenceTokenFilePath)).toBe(true);
         const users = readYML(sequenceTokenFilePath);
@@ -211,7 +213,7 @@ describe("saveSequenceTokenFile",  () => {
     });
 
     test("User not in user.yml",  () => {
-        const initUtilsObj = new initUtils();
+        const initUtilsObj = new initUtils(repoPath);
         fs.writeFileSync(sequenceTokenFilePath, yaml.safeDump(sequenceTokenFileData));
         initUtilsObj.saveSequenceTokenFile(ANOTHER_TEST_EMAIL);
         expect(fs.existsSync(sequenceTokenFilePath)).toBe(true);
@@ -252,7 +254,6 @@ describe("uploadRepo",  () => {
     let baseRepoPath;
     let repoPath;
     let syncIgnorePath;
-    let filePath;
     let configPath;
     let userFilePath;
     let sequenceTokenFilePath;
@@ -267,20 +268,18 @@ describe("uploadRepo",  () => {
         // Create directories
         fs.mkdirSync(baseRepoPath, {recursive: true});
         fs.mkdirSync(repoPath, {recursive: true});
+        
         untildify.mockReturnValue(baseRepoPath);
         createSystemDirectories();
-
+        writeTestRepoFiles(repoPath);
         configPath = getConfigFilePath(baseRepoPath);
         userFilePath = getUserFilePath(baseRepoPath);
         sequenceTokenFilePath = getSeqTokenFilePath(baseRepoPath);
         syncIgnorePath = getSyncIgnoreFilePath(repoPath);
-        filePath = path.join(repoPath, "file.js");
-    
         const configUtil = new Config(repoPath, configPath);
         configUtil.addRepo();
         fs.writeFileSync(configPath, yaml.safeDump(configData));
         fs.writeFileSync(syncIgnorePath, SYNC_IGNORE_DATA+"\nignore.js");
-        fs.writeFileSync(filePath, DUMMY_FILE_CONTENT);
         fs.writeFileSync(path.join(repoPath, "ignore.js"), DUMMY_FILE_CONTENT);
     });
 
@@ -293,8 +292,8 @@ describe("uploadRepo",  () => {
         // Generate ItemPaths
         const initUtilsObj = new initUtils(repoPath);
         const itemPaths = initUtilsObj.getSyncablePaths();
-        // 1 is .syncignore, other is file.js
-        expect(itemPaths).toHaveLength(2);
+        // Files count from TEST_RESPONSE_DATA
+        expect(itemPaths).toHaveLength(3);
         // Mock response for checkServerDown
         fetchMock.mockResponseOnce(JSON.stringify({status: false}));
 
@@ -305,7 +304,8 @@ describe("uploadRepo",  () => {
         const config = readYML(configPath);
         expect(config.repos[repoPath].branches[DEFAULT_BRANCH]).toStrictEqual({
             ".syncignore": null,
-            "file.js": null,
+            "file_1.js": null,
+            "directory/file_2.js": null,
         });
         expect(fetchMock).toHaveBeenCalledTimes(1);
     });
@@ -314,8 +314,8 @@ describe("uploadRepo",  () => {
         // Generate ItemPaths
         const initUtilsObj = new initUtils(repoPath);
         const itemPaths = initUtilsObj.getSyncablePaths();
-        // 1 is .syncignore, other is file.js
-        expect(itemPaths).toHaveLength(2);
+        // Files count from TEST_RESPONSE_DATA
+        expect(itemPaths).toHaveLength(3);
         // Mock response for checkServerDown
         fetchMock
             .mockResponseOnce(JSON.stringify({status: true}))
@@ -361,15 +361,25 @@ describe("uploadRepo",  () => {
         configUtil.removeRepo();
         const initUtilsObj = new initUtils(repoPath);
         const itemPaths = initUtilsObj.getSyncablePaths();
-        // 1 is .syncignore, other is file.js
-        expect(itemPaths).toHaveLength(2);
+        // Files count from TEST_RESPONSE_DATA
+        expect(itemPaths).toHaveLength(3);
         // Mock response for checkServerDown
         fetchMock
             .mockResponseOnce(JSON.stringify({status: true}))
             .mockResponseOnce(JSON.stringify(TEST_REPO_RESPONSE));
-    
+
+        const filePaths = itemPaths.map(itemPath => itemPath.file_path);
+        const pathUtilsObj = new pathUtils(repoPath, DEFAULT_BRANCH);
+        // copy files to .originals repo
+        const originalsRepoBranchPath = pathUtilsObj.getOriginalsRepoBranchPath();
+        initUtilsObj.copyFilesTo(filePaths, originalsRepoBranchPath);
+        // copy files to .shadow repo
+        const shadowRepoBranchPath = pathUtilsObj.getShadowRepoBranchPath();
+        initUtilsObj.copyFilesTo(filePaths, shadowRepoBranchPath);
+            
         await initUtilsObj.uploadRepo(DEFAULT_BRANCH, "ACCESS_TOKEN", itemPaths,
             TEST_EMAIL, false);
+        await waitFor(3);
 
         // Verify file Ids have been added in config
         const config = readYML(configPath);
@@ -383,7 +393,7 @@ describe("uploadRepo",  () => {
         users = readYML(userFilePath);
         expect(users[TEST_USER.email].access_key).toStrictEqual(TEST_USER.iam_access_key);
         expect(users[TEST_USER.email].secret_key).toStrictEqual(TEST_USER.iam_secret_key);
-        // Verify error msg
+        // Verify notification msg
         expect(vscode.window.showInformationMessage).toHaveBeenCalledTimes(1);
         expect(vscode.window.showInformationMessage.mock.calls[0][0]).toStrictEqual(NOTIFICATION.REPO_SYNCED);
         expect(vscode.window.showErrorMessage).toHaveBeenCalledTimes(0);
@@ -399,10 +409,10 @@ describe("uploadRepo",  () => {
         expect(body.name).toStrictEqual(path.basename(repoPath));
         expect(body.is_public).toBe(false);
         expect(body.branch).toStrictEqual(DEFAULT_BRANCH);
-        expect(body.source).toStrictEqual(DIFF_SOURCE);
+        expect(body.source).toStrictEqual(VSCODE);
         expect(body.platform).toStrictEqual(os.platform());
         const files_data = JSON.parse(body.files_data);
-        [".syncignore", "file.js"].forEach(key => {
+        Object.keys(TEST_REPO_RESPONSE.file_path_and_id).forEach(key => {
             expect(files_data[key]).toBeTruthy();
         });
     });
@@ -414,8 +424,8 @@ describe("uploadRepo",  () => {
 
         const initUtilsObj = new initUtils(repoPath);
         const itemPaths = initUtilsObj.getSyncablePaths();
-        // 1 is .syncignore, other is file.js
-        expect(itemPaths).toHaveLength(2);
+        // Files count from TEST_RESPONSE_DATA
+        expect(itemPaths).toHaveLength(3);
         // Mock response for checkServerDown
         fetchMock
             .mockResponseOnce(JSON.stringify({status: true}))